// generated/src/app/api-models/zip-codes.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ZipCodes {
  zip: string;
  city: string;
  state: string;
  county: string;
  areaCode: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  zipCodeQualifier: string;
}